[refactor] add_exam_page widget splitting into different widgets (#47)
